Mahdi - Add Discount And Rate In Shows - 8/6/1402

+++ BN_Project.Domain/Entities/OrderBasket/Order.cs
@@ -6,7 +6,7 @@ namespace BN_Project.Domain.Entities
 {
     public class Order : BaseEntity
     {
-        public long FinalPrice { get; set; }
+        public int FinalPrice { get; set; }
 
         public OrderStatus Status { get; set; }
